split Map.js into components + add comments for future devs
